+++ .repo/scratchad.sql
@@ -1,0 +1,9 @@
+-- PRAGMA table_list;
+-- PRAGMA table_info(Artifact);
+-- DELETE FROM Artifact;
+--
+
+SELECT *
+FROM Artifact
+ORDER BY `date` DESC
+LIMIT 10;
