+++ functions/cdx.js
@@ -1,7 +1,64 @@
-import { AuthResult, OSV, Server, ensureStrReqBody, hex, isSPDX, saveArtifact } from "@/utils";
+import { AuthResult, ensureStrReqBody, hex, isCDX, OSV, saveArtifact, Server } from "@/utils";
 import { PrismaD1 } from '@prisma/adapter-d1';
 import { PrismaClient } from '@prisma/client';
 
+
+export async function onRequestGet(context) {
+    const {
+        request, // same as existing Worker API
+        env, // same as existing Worker API
+        params, // if filename includes [id] or [[path]]
+        waitUntil, // same as ctx.waitUntil in existing Worker API
+        next, // used for middleware or to fetch assets
+        data, // arbitrary space for passing data between middlewares
+    } = context
+    const adapter = new PrismaD1(env.d1db)
+    const prisma = new PrismaClient({
+        adapter,
+        transactionOptions: {
+            maxWait: 1500, // default: 2000
+            timeout: 2000, // default: 5000
+        },
+    })
+    const verificationResult = await (new Server(request, prisma)).authenticate()
+    if (!verificationResult.isValid) {
+        return Response.json({ ok: false, result: verificationResult.message })
+    }
+    const { searchParams } = new URL(request.url)
+    const take = parseInt(searchParams.get('take'), 10) || 50
+    const skip = parseInt(searchParams.get('skip'), 10) || 0
+    let cdx = await prisma.CycloneDXInfo.findMany({
+        where: {
+            orgId: verificationResult.session.orgId,
+        },
+        omit: {
+            memberEmail: true,
+        },
+        include: {
+            repo: true,
+            artifact: {
+                include: {
+                    downloadLinks: true
+                }
+            },
+        },
+        take,
+        skip,
+        orderBy: {
+            createdAt: 'desc',
+        },
+    })
+    cdx = cdx.map(item => {
+        let updatedItem = { ...item }
+        if (item.artifact && item.artifact.downloadLinks && item.artifact.downloadLinks.length) {
+            updatedItem.downloadLink = item.artifact.downloadLinks.filter(l => l.contentType === 'application/vnd.cyclonedx+json')?.pop()?.url
+        }
+        delete updatedItem.artifact
+
+        return updatedItem
+    })
+    return Response.json({ ok: true, cdx })
+}
 
 export async function onRequestPost(context) {
     const {
@@ -24,70 +81,68 @@ export async function onRequestPost(context) {
     if (!verificationResult.isValid) {
         return Response.json({ ok: false, result: verificationResult.message })
     }
+
     const files = []
     let errors = new Set()
     try {
         const body = await ensureStrReqBody(request)
         const inputs = JSON.parse(body)
-        for (const spdx of inputs) {
-            if (!isSPDX(spdx)) {
-                return Response.json({ ok: false, error: { message: 'SPDX is missing necessary fields.' } })
+        for (const cdx of inputs) {
+            if (!isCDX(cdx)) {
+                return Response.json({ ok: false, error: { message: 'CDX is missing necessary fields.' } })
             }
-            const spdxId = await makeId(spdx)
-            const originalSpdx = await prisma.SPDXInfo.findFirst({
+            const componentsJSON = JSON.stringify(cdx.components)
+            const cdxId = await hex(cdx.metadata?.component?.name + componentsJSON)
+
+            const originalCdx = await prisma.CycloneDXInfo.findFirst({
                 where: {
-                    spdxId,
+                    cdxId,
                     orgId: verificationResult.session.orgId,
                 }
             })
-            let artifact;
-            if (!originalSpdx) {
-                const spdxStr = JSON.stringify(spdx)
-                artifact = await saveArtifact(prisma, env.r2artifacts, spdxStr, crypto.randomUUID(), `spdx`)
-            }
-            const artifactUuid = originalSpdx?.artifactUuid || artifact?.uuid
-            const spdxData = {
-                spdxId,
+
+            const artifactUuid = originalCdx?.artifactUuid || cdx.serialNumber.startsWith('urn:uuid:') ? cdx.serialNumber.substring(9) : crypto.randomUUID()
+            const cdxStr = JSON.stringify(cdx)
+            const artifact = await saveArtifact(prisma, env.r2artifacts, cdxStr, artifactUuid, `cyclonedx`)
+            const cdxData = {
+                cdxId,
                 artifactUuid,
                 source: 'upload',
                 orgId: verificationResult.session.orgId,
                 memberEmail: verificationResult.session.memberEmail,
-                repoName: '',
-                spdxVersion: spdx.spdxVersion,
-                dataLicense: spdx.dataLicense,
-                name: spdx.name,
-                documentNamespace: spdx.documentNamespace,
-                createdAt: (new Date(spdx.creationInfo.created)).getTime(),
-                toolName: spdx.creationInfo.creators.join(', '),
-                documentDescribes: spdx.documentDescribes.join(','),
-                packagesCount: spdx.packages.length,
-                comment: spdx.creationInfo?.comment || '',
+                cdxVersion: cdx.specVersion,
+                serialNumber: cdx.serialNumber,
+                name: cdx.metadata?.component?.name,
+                version: cdx.metadata?.component?.version,
+                createdAt: (new Date(cdx.metadata.timestamp)).getTime(),
+                toolName: cdx.metadata.tools.map(t => `${t?.vendor} ${t?.name} ${t?.version}`.trim()).join(', '),
+                externalReferencesCount: cdx.metadata.component?.externalReferences?.length || 0,
+                componentsCount: cdx.components?.length || 0,
+                dependenciesCount: cdx.dependencies?.length || 0,
             }
-            const info = await prisma.SPDXInfo.upsert({
+            const info = await prisma.CycloneDXInfo.upsert({
                 where: {
-                    spdxId,
-                    orgId: verificationResult.session.orgId,
+                    cdxId,
+                    memberEmail: verificationResult.session.memberEmail,
                 },
                 update: {
-                    createdAt: spdxData.createdAt,
-                    comment: spdxData.comment
+                    createdAt: cdxData.createdAt,
+                    serialNumber: cdxData.serialNumber
                 },
-                create: spdxData
+                create: cdxData
             })
-            console.log(`/github/repos/spdx ${spdxId} kid=${verificationResult.session.kid}`, info)
-            files.push(spdxData)
+            console.log(`/upload/cdx ${cdxId} kid=${verificationResult.session.kid}`, info)
+            files.push(cdxData)
 
-            const osvQueries = spdx.packages.flatMap(pkg => {
-                if (!pkg?.externalRefs) { return }
-                return pkg.externalRefs
-                    .filter(ref => ref?.referenceType === 'purl')
-                    .map(ref => ({
-                        referenceLocator: decodeURIComponent(ref.referenceLocator),
-                        name: pkg.name,
-                        version: pkg?.versionInfo,
-                        license: pkg?.licenseConcluded || pkg?.licenseDeclared,
-                    }))
-            }).filter(q => q?.referenceLocator)
+            const osvQueries = cdx.components.map(component => {
+                if (!component?.purl) { return }
+                return {
+                    referenceLocator: decodeURIComponent(component.purl),
+                    name: component.name,
+                    version: component?.version,
+                    license: component?.licenses?.map(l => l.license?.id || '').join(' '),
+                }
+            })
             const osv = new OSV()
             const queries = osvQueries.map(q => ({ package: { purl: q?.referenceLocator } }))
             const results = await osv.queryBatch(prisma, verificationResult.session.orgId, verificationResult.session.memberEmail, queries)
@@ -113,7 +168,7 @@ export async function onRequestPost(context) {
                         packageVersion: version,
                         packageLicense: license,
                         maliciousSource: vuln.id.startsWith("MAL-"),
-                        spdxId
+                        cdxId
                     }
                     const originalFinding = await prisma.Finding.findFirst({
                         where: {
@@ -130,14 +185,15 @@ export async function onRequestPost(context) {
                                 uuid: originalFinding.uuid,
                             },
                             data: {
-                                spdxId,
+                                cdxId,
                                 modifiedAt: findingData.modifiedAt
                             },
                         })
                     } else {
                         finding = await prisma.Finding.create({ data: findingData })
                     }
-                    // console.log(`findings SCA`, finding)
+                    console.log(`findings SCA`, finding)
+                    // TODO lookup EPSS
                     const vexData = {
                         findingUuid: finding.uuid,
                         createdAt: (new Date()).getTime(),
@@ -164,7 +220,7 @@ export async function onRequestPost(context) {
                     } else {
                         vex = await prisma.Triage.create({ data: vexData })
                     }
-                    // console.log(`findings VEX`, vex)
+                    console.log(`findings VEX`, vex)
                 }
                 i++
             }
@@ -181,9 +237,4 @@ export async function onRequestPost(context) {
     }
 
     return Response.json({ ok: true, files, error: { message: errors } })
-}
-
-const makeId = async spdx => {
-    const packages = JSON.stringify(spdx.packages)
-    return hex(spdx.name + packages)
 }
