--- src/pages/SARIFManager.vue
@@ -1,758 +1,0 @@
-<script setup>
-import router from "@/router";
-import { Client, isJSON, isSARIF, timeAgo } from '@/utils';
-import { reactive } from 'vue';
-import { useTheme } from 'vuetify';
-
-const client = new Client()
-const { global } = useTheme()
-
-const initialState = {
-    uploadError: "",
-    uploadSuccess: "",
-    error: "",
-    warning: "",
-    success: "",
-    info: "",
-    loading: false,
-    files: [],
-    uploads: [],
-    github: [],
-}
-const delArtifact = ref()
-
-const state = reactive({
-    ...initialState,
-})
-class Controller {
-    constructor() {
-        this.refresh(true)
-    }
-
-    async refresh(initial = false) {
-        clearAlerts()
-        state.loading = true
-        try {
-            const pageSize = 50
-            let hasMore = true
-            let skip = 0
-            while (hasMore) {
-                const { data } = await client.get(`/sarif/results?take=${pageSize}&skip=${skip}`)
-                if (data.ok) {
-                    if (data?.sarif) {
-                        data.sarif.filter(item => item.source === "upload").forEach(sarif => state.uploads.push(sarif))
-                        data.sarif.filter(item => item.source === "GitHub").forEach(sarif => state.github.push(sarif))
-                    }
-                } else if (typeof data === "string" && !isJSON(data)) {
-                    if (initial !== true) {
-                        state.warning = "SARIF data could not be retrieved, please try again later."
-                    }
-                    break
-                } else if (data?.error?.message) {
-                    state.loading = false
-                    state.error = data.error.message
-                    return
-                } else if (["Expired", "Revoked", "Forbidden"].includes(data?.result)) {
-                    state.loading = false
-                    state.info = data.result
-                    setTimeout(() => router.push('/logout'), 2000)
-                    return
-                } else {
-                    if (initial !== true) {
-                        state.info = "No SARIF data available."
-                    }
-                    break
-                }
-                if (data.sarif.length < pageSize) {
-                    hasMore = false
-                    if (initial !== true) {
-                        state.info = "Refreshed SARIF"
-                    }
-                } else {
-                    skip += pageSize
-                }
-            }
-            state.loading = false
-
-            return
-        } catch (e) {
-            console.error(e)
-            state.error = `${e.code} ${e.message}`
-            state.loading = false
-        }
-        state.warning = "No SARIF data available."
-    }
-
-    async upload() {
-        clearAlerts()
-        state.loading = true
-        const files = []
-        try {
-            for (const blob of state.files) {
-                const text = await blob.text()
-                if (!isJSON(text)) {
-                    state.uploadError = "Provided file does not contain a JSON string."
-                    break
-                }
-                const json = JSON.parse(text)
-                if (isSARIF(json)) {
-                    files.push(json)
-                } else {
-                    state.uploadError = "Provided file does not contain valid SARIF."
-                    break
-                }
-            }
-            if (!files.length) {
-                state.uploadError = "No SARIF files were provided."
-                return
-            }
-            console.log(files)
-            const { data } = await client.post(`/sarif/upload`, files)
-            state.loading = false
-
-            if (typeof data === "string" && !isJSON(data)) {
-                state.uploadError = "SARIF data could not be uploaded, please try again later."
-
-                return
-            }
-            if (data?.error?.message) {
-                state.uploadError = data?.error?.message
-            }
-            if (["Expired", "Revoked", "Forbidden"].includes(data?.result)) {
-                state.uploadError = data.result
-
-                return setTimeout(() => router.push('/logout'), 2000)
-            }
-            if (!data?.sarif?.length) {
-                state.uploadError = "No SARIF data available."
-            } else {
-                for (const file of data.sarif) {
-                    if (file.source === "upload" && !state.uploads.some(f => f.reportId === file.reportId)) {
-                        state.uploads.push(file)
-                    }
-                    if (file.source === "GitHub" && !state.github.some(f => f.reportId === file.reportId)) {
-                        state.github.push(file)
-                    }
-                }
-                state.uploadSuccess = "Uploaded SARIF, you may close this dialogue now."
-            }
-
-            return
-        } catch (e) {
-            console.error(e)
-            state.uploadError = typeof e === "string" ? e : `${e.code} ${e.message}`
-            state.loading = false
-        }
-    }
-    deleteArtifact = async (record, isActive) => {
-        clearAlerts()
-        state.loading = true
-        try {
-            const { data } = await client.delete(`/sarif/${record.reportId}`)
-            state.loading = false
-
-            if (!data?.ok) {
-                state.error = data?.error?.message || 'Server error, please try again later.'
-                isActive.value = false
-
-                return
-            }
-            if (["Expired", "Revoked", "Forbidden"].includes(data?.result)) {
-                state.info = data.result
-                isActive.value = false
-
-                return setTimeout(() => router.push('/logout'), 2000)
-            }
-            if (typeof data === "string" && !isJSON(data)) {
-                state.error = "Data could not be deleted, please try again later."
-                isActive.value = false
-
-                return
-            }
-            if (data) {
-                if (record.source === "upload") {
-                    state.uploads = state.uploads.filter(o => o.reportId !== record.reportId)
-                } else if (record.source === "GitHub") {
-                    state.github = state.github.filter(o => o.reportId !== record.reportId)
-                }
-                state.success = "Artifact deleted successfully."
-            } else {
-                state.info = data?.result || 'No change'
-            }
-            isActive.value = false
-
-            return
-        } catch (e) {
-            console.error(e)
-            state.error = typeof e === "string" ? e : `${e.code} ${e.message}`
-            state.loading = false
-            isActive.value = false
-        }
-    }
-    groupedByOrg = () => {
-        return state.github.reduce((acc, sarif) => {
-            const [orgName, repoName] = sarif.fullName.split('/')
-            let group = acc.find(group => group.orgName === orgName)
-
-            if (!group) {
-                group = {
-                    orgName: orgName,
-                    downloadLink: sarif?.downloadLink,
-                    avatarUrl: sarif?.repo?.avatarUrl,
-                    sarif: []
-                };
-                acc.push(group)
-            }
-            group.sarif.push({ ...sarif, orgName, repoName })
-            return acc
-        }, [])
-    }
-}
-
-function clearAlerts() {
-    state.uploadError = ''
-    state.uploadSuccess = ''
-    state.error = ''
-    state.warning = ''
-    state.success = ''
-    state.info = ''
-}
-
-const controller = reactive(new Controller())
-</script>
-
-<template>
-    <VRow>
-        <VCol cols="12">
-            <VAlert
-                v-if="state.error"
-                color="error"
-                icon="$error"
-                title="Error"
-                :text="state.error"
-                border="start"
-                variant="tonal"
-            />
-            <VAlert
-                v-if="state.warning"
-                color="warning"
-                icon="$warning"
-                title="Warning"
-                :text="state.warning"
-                border="start"
-                variant="tonal"
-            />
-            <VAlert
-                v-if="state.success"
-                color="success"
-                icon="$success"
-                title="Success"
-                :text="state.success"
-                border="start"
-                variant="tonal"
-            />
-            <VAlert
-                v-if="state.info"
-                color="info"
-                icon="$info"
-                title="Information"
-                :text="state.info"
-                border="start"
-                variant="tonal"
-            />
-        </VCol>
-
-        <VCol cols="12">
-            <VBtn
-                text="Refresh"
-                prepend-icon="mdi-refresh"
-                variant="text"
-                :color="global.name.value === 'dark' ? '#fff' : '#272727'"
-                :disabled="state.loading"
-                @click="controller.refresh"
-            />
-            <VDialog
-                width="50%"
-                persistent
-            >
-                <template v-slot:activator="{ props: activatorProps }">
-                    <VBtn
-                        class="text-right"
-                        prepend-icon="mdi-upload"
-                        text="Upload SARIF"
-                        variant="text"
-                        :color="global.name.value === 'dark' ? '#fff' : '#272727'"
-                        v-bind="activatorProps"
-                    />
-                </template>
-                <template v-slot:default="{ isActive }">
-                    <VCard title="Select SARIF files">
-                        <VProgressLinear
-                            :active="state.loading"
-                            :indeterminate="state.loading"
-                            color="primary"
-                            absolute
-                            bottom
-                        ></VProgressLinear>
-                        <VAlert
-                            v-if="state.uploadError"
-                            color="error"
-                            icon="$error"
-                            title="Error"
-                            :text="state.uploadError"
-                            border="start"
-                            variant="tonal"
-                        />
-                        <VAlert
-                            v-if="state.uploadSuccess"
-                            color="success"
-                            icon="$success"
-                            title="Success"
-                            :text="state.uploadSuccess"
-                            border="start"
-                            variant="tonal"
-                        />
-                        <VDivider class="mt-3"></VDivider>
-                        <VCardText class="px-4">
-                            <VFileInput
-                                v-model="state.files"
-                                :show-size="1000"
-                                accept="application/json"
-                                label="File input"
-                                placeholder="Select your files"
-                                prepend-icon="mdi-paperclip"
-                                variant="outlined"
-                                counter
-                                multiple
-                            >
-                                <template v-slot:selection="{ fileNames }">
-                                    <template
-                                        v-for="(fileName, index) in fileNames"
-                                        :key="fileName"
-                                    >
-                                        <VChip
-                                            v-if="index < 2"
-                                            class="me-2"
-                                            size="small"
-                                            label
-                                        >
-                                            {{ fileName }}
-                                        </VChip>
-
-                                        <span
-                                            v-else-if="index === 2"
-                                            class="text-overline text-grey-darken-3 mx-2"
-                                        >
-                                            +{{ state.files.length - 2 }} File(s)
-                                        </span>
-                                    </template>
-                                </template>
-                            </VFileInput>
-                        </VCardText>
-                        <VDivider></VDivider>
-                        <VCardActions class="mt-3">
-                            <VBtn
-                                text="Close"
-                                @click="isActive.value = false"
-                            />
-                            <VSpacer></VSpacer>
-                            <VBtn
-                                text="Upload"
-                                variant="flat"
-                                @click="controller.upload"
-                            />
-                        </VCardActions>
-                    </VCard>
-                </template>
-            </VDialog>
-        </VCol>
-
-        <VCol
-            cols="12"
-            v-if="state.github.length || state.loading"
-        >
-            <VCard
-                title="Github"
-                prepend-icon="line-md:github-loop"
-            >
-                <VExpansionPanels accordion>
-                    <VExpansionPanel
-                        v-for="(group, k) in controller.groupedByOrg()"
-                        :key="k"
-                    >
-                        <VExpansionPanelTitle
-                            v-if="group.sarif.length"
-                            class="text-subtitle-1"
-                            v-slot="{ expanded }"
-                        >
-                            <VRow no-gutters>
-                                <VCol
-                                    class="d-flex justify-start"
-                                    cols="12"
-                                >
-                                    <img
-                                        :src="group.avatarUrl"
-                                        width="25"
-                                        height="25"
-                                        class="me-3"
-                                    >
-                                    <span>{{ group.orgName }}&nbsp;
-                                        <span v-if="!expanded">({{ group.sarif.length }} results)</span>
-                                    </span>
-                                </VCol>
-                            </VRow>
-                        </VExpansionPanelTitle>
-                        <VExpansionPanelText>
-                            <VSkeletonLoader
-                                v-if="state.loading"
-                                type="table-row@10"
-                            />
-                            <VTable
-                                v-else
-                                height="20rem"
-                                fixed-header
-                            >
-                                <thead>
-                                    <tr>
-                                        <th class="text-uppercase">
-                                            Repository
-                                        </th>
-                                        <th class="text-center">
-                                            Git Reference
-                                        </th>
-                                        <th class="text-center">
-                                            Findings
-                                        </th>
-                                        <th class="text-center">
-                                            Rules Count
-                                        </th>
-                                        <th class="text-center">
-                                            Tool
-                                        </th>
-                                        <th class="text-center">
-                                            Commit SHA
-                                        </th>
-                                        <th class="text-center">
-                                            Reports
-                                        </th>
-                                        <th class="text-center">
-                                            Created Date
-                                        </th>
-                                        <th class="text-right">
-                                        </th>
-                                    </tr>
-                                </thead>
-
-                                <tbody>
-                                    <tr
-                                        v-for="(result, i) in group.sarif"
-                                        :key="i"
-                                    >
-                                        <td>
-                                            {{ result.repoName }}
-                                        </td>
-                                        <td class="text-center">
-                                            {{ result.ref }}
-                                        </td>
-                                        <td class="text-center">
-                                            {{ result.resultsCount }}
-                                        </td>
-                                        <td class="text-center">
-                                            {{ result.rulesCount }}
-                                        </td>
-                                        <td class="text-center">
-                                            {{ result.toolName }} {{ result.toolVersion }}
-                                        </td>
-                                        <td class="text-center">
-                                            {{ result.commitSha }}
-                                        </td>
-                                        <td class="text-center">
-                                            {{ result.results.length }}
-                                        </td>
-                                        <td class="text-center">
-                                            <VTooltip
-                                                :text="(new Date(result.createdAt)).toLocaleString()"
-                                                location="left"
-                                            >
-                                                <template v-slot:activator="{ props }">
-                                                    <time
-                                                        v-bind="props"
-                                                        :datetime="(new Date(result.createdAt)).toISOString()"
-                                                    >
-                                                        {{ timeAgo(new Date(result.createdAt)) }}
-                                                    </time>
-                                                </template>
-                                            </VTooltip>
-                                        </td>
-                                        <td class="text-right">
-                                            <VTooltip
-                                                v-if="result.downloadLink"
-                                                text="Download"
-                                                location="left"
-                                            >
-                                                <template v-slot:activator="{ props }">
-                                                    <VBtn
-                                                        v-bind="props"
-                                                        variant="plain"
-                                                        icon="line-md:cloud-alt-download-filled-loop"
-                                                        density="comfortable"
-                                                        color="secondary"
-                                                        target="_blank"
-                                                        :title="result.downloadLink"
-                                                        :href="result.downloadLink"
-                                                    >
-                                                    </VBtn>
-                                                </template>
-                                            </VTooltip>
-                                            <VDialog
-                                                persistent
-                                                :activator="`#sarif${k}_${i}`"
-                                                max-width="340"
-                                            >
-                                                <template v-slot:default="{ isActive }">
-                                                    <VCard
-                                                        prepend-icon="weui:delete-on-filled"
-                                                        text="This will delete the SARIF Artifact permanantly, it cannot be undone."
-                                                        title="Delete Artifact?"
-                                                    >
-                                                        <template v-slot:actions>
-                                                            <VBtn
-                                                                class="ml-auto"
-                                                                text="Close"
-                                                                @click="isActive.value = false"
-                                                            ></VBtn>
-                                                            <VBtn
-                                                                color="error"
-                                                                variant="tonal"
-                                                                text="Delete"
-                                                                @click="controller.deleteArtifact(result, isActive)"
-                                                            ></VBtn>
-                                                        </template>
-                                                    </VCard>
-                                                </template>
-                                            </VDialog>
-                                            <VTooltip
-                                                text="Delete"
-                                                location="left"
-                                            >
-                                                <template v-slot:activator="{ props }">
-                                                    <VBtn
-                                                        v-bind="props"
-                                                        :id="`sarif${k}_${i}`"
-                                                        variant="plain"
-                                                        icon="weui:delete-on-filled"
-                                                        density="comfortable"
-                                                        color="error"
-                                                    >
-                                                    </VBtn>
-                                                </template>
-                                            </VTooltip>
-                                        </td>
-                                    </tr>
-                                </tbody>
-                            </VTable>
-                        </VExpansionPanelText>
-                    </VExpansionPanel>
-                </VExpansionPanels>
-            </VCard>
-        </VCol>
-        <VCol
-            cols="12"
-            v-if="state.uploads.length || state.loading"
-        >
-            <VCard
-                title="Uploads"
-                prepend-icon="uil-file-upload"
-            >
-                <VExpansionPanels accordion>
-                    <VExpansionPanel
-                        v-for="(sarif, i) in state.uploads"
-                        :key="i"
-                    >
-                        <VExpansionPanelTitle
-                            v-slot="{ expanded }"
-                            class="text-subtitle-1"
-                            v-if="sarif.results.length"
-                        >
-                            <VRow no-gutters>
-                                <VCol
-                                    class="d-flex justify-start"
-                                    cols="10"
-                                >
-                                    <img
-                                        src="/sarif-logo.png"
-                                        width="25"
-                                        height="25"
-                                        class="mr-2"
-                                    >
-                                    <span>{{ sarif.sarifId }}.json&nbsp;
-                                        <span v-if="!expanded">({{ sarif.results.length }} results)</span>
-                                    </span>
-                                </VCol>
-                                <VCol
-                                    class="text--secondary"
-                                    cols="2"
-                                >
-                                    <VFadeTransition leave-absolute>
-                                        <VRow
-                                            style="width: 100%"
-                                            no-gutters
-                                        >
-                                            <VCol
-                                                class="d-flex justify-end"
-                                                cols="12"
-                                            >
-                                                <VTooltip
-                                                    v-if="sarif.downloadLink"
-                                                    text="Download"
-                                                    location="left"
-                                                >
-                                                    <template v-slot:activator="{ props }">
-                                                        <VBtn
-                                                            v-bind="props"
-                                                            variant="plain"
-                                                            icon="line-md:cloud-alt-download-filled-loop"
-                                                            density="comfortable"
-                                                            color="secondary"
-                                                            target="_blank"
-                                                            :title="sarif.downloadLink"
-                                                            :href="sarif.downloadLink"
-                                                        >
-                                                        </VBtn>
-                                                    </template>
-                                                </VTooltip>
-                                                <VDialog
-                                                    persistent
-                                                    :activator="`#sarifUpload${i}`"
-                                                    max-width="340"
-                                                >
-                                                    <template v-slot:default="{ isActive }">
-                                                        <VCard
-                                                            prepend-icon="weui:delete-on-filled"
-                                                            text="This will delete the SARIF Artifact permanantly, it cannot be undone."
-                                                            title="Delete Artifact?"
-                                                        >
-                                                            <template v-slot:actions>
-                                                                <VBtn
-                                                                    class="ml-auto"
-                                                                    text="Close"
-                                                                    @click="isActive.value = false"
-                                                                ></VBtn>
-                                                                <VBtn
-                                                                    color="error"
-                                                                    variant="tonal"
-                                                                    text="Delete"
-                                                                    @click="controller.deleteArtifact(sarif, isActive)"
-                                                                ></VBtn>
-                                                            </template>
-                                                        </VCard>
-                                                    </template>
-                                                </VDialog>
-                                                <VTooltip
-                                                    text="Delete"
-                                                    location="left"
-                                                >
-                                                    <template v-slot:activator="{ props }">
-                                                        <VBtn
-                                                            v-bind="props"
-                                                            :id="`sarifUpload${i}`"
-                                                            variant="plain"
-                                                            icon="weui:delete-on-filled"
-                                                            density="comfortable"
-                                                            color="error"
-                                                        >
-                                                        </VBtn>
-                                                    </template>
-                                                </VTooltip>
-                                            </VCol>
-                                        </VRow>
-                                    </VFadeTransition>
-                                </VCol>
-                            </VRow>
-                        </VExpansionPanelTitle>
-                        <VExpansionPanelText v-if="sarif.results.length">
-                            <VSkeletonLoader
-                                v-if="state.loading"
-                                type="table-row@10"
-                            />
-                            <VTable
-                                v-else
-                                height="20rem"
-                                fixed-header
-                            >
-                                <thead>
-                                    <tr>
-                                        <th class="text-uppercase">
-                                            Rule
-                                        </th>
-                                        <th>
-                                            Level
-                                        </th>
-                                        <th>
-                                            Precision
-                                        </th>
-                                        <th>
-                                            Tool
-                                        </th>
-                                        <th>
-                                            SARIF Identifier
-                                        </th>
-                                        <th>
-                                            Message Text
-                                        </th>
-                                        <th>
-                                            Created Date
-                                        </th>
-                                    </tr>
-                                </thead>
-
-                                <tbody>
-                                    <tr
-                                        v-for="(result, i) in sarif.results"
-                                        :key="i"
-                                    >
-                                        <td>
-                                            {{ result.ruleId }}
-                                        </td>
-                                        <td class="text-center">
-                                            {{ result.level }}
-                                        </td>
-                                        <td class="text-center">
-                                            {{ result.precision }}
-                                        </td>
-                                        <td class="text-center">
-                                            {{ sarif.toolName }} {{ sarif.toolVersion }}
-                                        </td>
-                                        <td class="text-center">
-                                            {{ sarif.sarifId }}
-                                        </td>
-                                        <td class="text-center">
-                                            {{ result.messageText }}
-                                        </td>
-                                        <td class="text-center">
-                                            <VTooltip
-                                                :text="(new Date(sarif.createdAt)).toLocaleString()"
-                                                location="left"
-                                            >
-                                                <template v-slot:activator="{ props }">
-                                                    <time
-                                                        v-bind="props"
-                                                        :datetime="(new Date(sarif.createdAt)).toISOString()"
-                                                    >
-                                                        {{ timeAgo(new Date(sarif.createdAt)) }}
-                                                    </time>
-                                                </template>
-                                            </VTooltip>
-                                        </td>
-                                    </tr>
-                                </tbody>
-                            </VTable>
-                        </VExpansionPanelText>
-                    </VExpansionPanel>
-                </VExpansionPanels>
-            </VCard>
-
-        </VCol>
-
-    </VRow>
-</template>
