--- src/pages/VDRManager.vue
@@ -1,5 +1,0 @@
-<script setup>
-</script>
-
-<template>
-</template>
