--- src/pages/VEXManager.vue
@@ -1,5 +1,0 @@
-<script setup>
-</script>
-
-<template>
-</template>
